polish environment configuration
