+++ evaluators/eval_codellama_wyk.py
@@ -4,8 +4,7 @@ import torch
 import logging
 import argparse
 import warnings
-# import os
-# os.environ['CUDA_VISIBLE_DEVICES']='0,1,2,3'
+
 from pathlib import Path
 from datasets import load_dataset
 from transformers import AutoTokenizer, AutoModelForCausalLM
@@ -224,8 +223,8 @@ def main():
     dataset.cleanup_cache_files()  # for multiple evaluation
     print(dataset)
     if 'code_optimization' in args.data_load_name:
-        mem_save_path = Path(__file__).parent.parent / Path('results') / Path(f"mem_{args.result_save_name}")
-        time_save_path = Path(__file__).parent.parent / Path('results') / Path(f"time_{args.result_save_name}")
+        mem_save_path = Path(__file__).parent.parent / Path('results') / Path('raw') / Path(f"mem_{args.result_save_name}")
+        time_save_path = Path(__file__).parent.parent / Path('results') / Path('raw') / Path(f"time_{args.result_save_name}")
         logging.info("=====start mem optimiing=====")
         mem_ds = dataset.map(add_mem_optimization)
         mem_ds.to_json(mem_save_path, lines=True)
@@ -233,7 +232,7 @@ def main():
         time_ds = dataset.map(add_time_optimization)
         time_ds.to_json(time_save_path, lines=True)
     elif "code_summarization" in args.data_load_name:
-        save_path = Path(__file__).parent.parent / Path('results') / Path(args.result_save_name)
+        save_path = Path(__file__).parent.parent / Path('results') / Path('raw') / Path(args.result_save_name)
         dataset = dataset.map(add_code_summ)
         dataset.to_json(save_path, lines=True)
 
@@ -282,7 +281,6 @@ if __name__ == '__main__':
     temperature = args.temperature
     max_input_tokens = tokenizer.model_max_length  # 1000000000000000019884624838656
     # The maximum numbers of tokens to generate, ignoring the number of tokens in the prompt.
-    max_new_tokens = 2048  # max_output_tokens
+    max_new_tokens = 2048
 
     main()
-    # python scripts/eval_codellama.py
