+++ evaluators/eval_wizardcoder_wyk.py
@@ -4,8 +4,7 @@ import torch
 import logging
 import argparse
 import warnings
-# import os
-# os.environ['CUDA_VISIBLE_DEVICES']='4,7'
+
 from pathlib import Path
 from datasets import load_dataset
 from transformers import AutoTokenizer, AutoModelForCausalLM
@@ -239,8 +238,8 @@ def main():
     dataset.cleanup_cache_files()  # for multiple evaluation
     print(dataset)
     if 'code_optimization' in args.data_load_name:
-        mem_save_path = Path(__file__).parent.parent / Path('results') / Path(f"mem_{args.result_save_name}")
-        time_save_path = Path(__file__).parent.parent / Path('results') / Path(f"time_{args.result_save_name}")
+        mem_save_path = Path(__file__).parent.parent / Path('results') / Path('raw') / Path(f"mem_{args.result_save_name}")
+        time_save_path = Path(__file__).parent.parent / Path('results') / Path('raw') / Path(f"time_{args.result_save_name}")
         logging.info("=====start mem optimiing=====")
         mem_ds = dataset.map(add_mem_optimization)
         mem_ds.to_json(mem_save_path, lines=True)
@@ -248,7 +247,7 @@ def main():
         time_ds = dataset.map(add_time_optimization)
         time_ds.to_json(time_save_path, lines=True)
     elif "code_summarization" in args.data_load_name:
-        save_path = Path(__file__).parent.parent / Path('results') / Path(args.result_save_name)
+        save_path = Path(__file__).parent.parent / Path('results') / Path('raw') / Path(args.result_save_name)
         dataset = dataset.map(add_code_summ)
         dataset.to_json(save_path, lines=True)
 
@@ -300,5 +299,3 @@ if __name__ == '__main__':
     max_new_tokens = 2048
 
     main()
-    # python scripts/eval_wizardcoder.py --checkpoint /home/wyk/hf_cache/WizardCoder --data_load_name code_summarization_dataset_with_gt.jsonl --result_save_name code_summ_inference_wizardcoder.jsonl --log_file_name code_summ_inference_wizardcoder.log
-
